add clear module ABI & API

+++ crates/datastore/src/locking_tx_datastore/mut_tx.rs
@@ -3084,11 +3084,11 @@ impl MutTxId {
     }
 
     // Clears the table for `table_id`, removing all rows.
-    pub fn clear_table(&mut self, table_id: TableId) -> Result<usize> {
+    pub fn clear_table(&mut self, table_id: TableId) -> Result<u64> {
         // Get the commit table.
         let (commit_table, commit_bs, ..) = self.committed_state_write_lock.get_table_and_blob_store(table_id)?;
 
-        // Get the insert table and delete all rows from it.
+        // Get the insert table and delete all rows from it.s
         let (tx_table, tx_blob_store, delete_table) = self
             .tx_state
             .get_table_and_blob_store_or_create_from(table_id, commit_table);
